requirements.txt
aiohttp==3.6.2
apipkg==1.5
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==3.1.1
discord-webhook==0.9.0
discord.py==1.5.0
execnet==1.7.1
idna==2.10
iniconfig==1.0.1
mock==4.0.2
multidict==4.7.6
mysql-connector-python==8.0.21
packaging==20.4
paramiko==2.7.2
path==15.0.0
path.py==12.5.0
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
requests==2.24.0
scp==0.13.2
six==1.15.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
yarl==1.6.0
